refactor: move the canvas grid, chat-card, and intent-bus hooks out of page.tsx
Refs #1936.

Page line count changed from 4,467 to 3,513 lines.
